Significant changes to code but not functionality in this commit. Initial intention was to fix the IBU issue in Bug 3138339 (SourceForge) plus some other IBU calc issues but one thing led to another and numerous files needed to be updated. Code in many of the files for this project is nearly impossible to read due to poor/no formatting. In most of the updated files in this commit:
* Major re-formatting of code to increase readability.

* Fixes to logic errors; Conditional branches that will never execute 
under any circumstance, for example.

* Implemented PHP best practices in code style. The standards listed on 
the PEAR site are a good start: 
http://pear.php.net/manual/en/standards.php

* Implemented arrays, hashes, constants and 'else' conditionals where 
relevant. These features of the language seem to have been overlooked
in most of the files updated.

* All bitterness calculations now use the admin/lib/bitterness.lib.php 
library. Cut-n-paste instances of the formulas have been removed from
all modules in the project now (unless I missed any).

* admin/includes/constants.inc.php has been added to the project to 
contain constants. Many constants can probably be moved into this file 
over time; However, those containing relative paths may prove to be 
difficult so they might be better left out.

* Major changes to the method of collecting POST data and formatting it
for commits to the database in the admin/process.php module. The
functionality is the same but it's much easier to read and manage now.
Many of the sections in this file still need to be updated, however.

* Trying to add comment headers to each file as I think about it. A
short description of what the file does goes a long way in helping
to understand the flow.

Overall, there was a reduction of over 2,900 lines of code from these
files being committed. The break down on the larger ones:

recipe_calculator.php = 146
process.php           = 2066
importFormVar         = 139
calcFormVar           = 143
calculations          = 414

I hope I haven't stepped on any toes with this update but I believe the
changes make it easier for everyone to read and follow the code and,
therefore, easier to fix bugs, add features, etc.